fix err

+++ src/expr.cpp
@@ -176,7 +176,7 @@ AST *Parser::expr_primary() {
       std::string _; 
       type = read_declarator(_, type);
       token.expect_skip(")");
-      return new NumberAST(data_layout->getTypeAllocSize(type));
+      return new NumberAST((int)data_layout->getTypeAllocSize(type));
     } else {
       AST *expr = expr_entry();
       token.expect_skip(")");
